Do not get semantic model to ask syntaqctic question.
