+++ src/Features/CSharp/Portable/Completion/CompletionProviders/ExplicitInterfaceMemberCompletionProvider.cs
@@ -58,14 +58,13 @@ namespace Microsoft.CodeAnalysis.CSharp.Completion.Providers
                 var position = context.Position;
                 var cancellationToken = context.CancellationToken;
 
-                var semanticModel = await document.ReuseExistingSpeculativeModelAsync(position, cancellationToken).ConfigureAwait(false);
-                var syntaxTree = semanticModel.SyntaxTree;
+                var syntaxTree = await document.GetRequiredSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
 
                 var syntaxFacts = document.GetRequiredLanguageService<ISyntaxFactsService>();
                 var semanticFacts = document.GetRequiredLanguageService<ISemanticFactsService>();
 
                 if (syntaxFacts.IsInNonUserCode(syntaxTree, position, cancellationToken) ||
-                    semanticFacts.IsPreProcessorDirectiveContext(semanticModel, position, cancellationToken))
+                    syntaxFacts.IsPreProcessorDirectiveContext(syntaxTree, position, cancellationToken))
                 {
                     return;
                 }
@@ -83,6 +82,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Completion.Providers
                 // Bind the interface name which is to the left of the dot
                 var name = specifierNode.Name;
 
+                var semanticModel = await document.ReuseExistingSpeculativeModelAsync(position, cancellationToken).ConfigureAwait(false);
                 var symbol = semanticModel.GetSymbolInfo(name, cancellationToken).Symbol as ITypeSymbol;
                 if (symbol?.TypeKind != TypeKind.Interface)
                     return;
