+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/CSharp/Extensions/ContextQuery/SyntaxTreeExtensions.cs
@@ -1885,26 +1885,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions.ContextQuery
                 !syntaxTree.IsConstantExpressionContext(position, tokenOnLeftOfPosition);
         }
 
-        public static bool IsPreProcessorDirectiveContext(this SyntaxTree syntaxTree, int position, SyntaxToken preProcessorTokenOnLeftOfPosition, CancellationToken cancellationToken)
-        {
-            var token = preProcessorTokenOnLeftOfPosition;
-            var directive = token.GetAncestor<DirectiveTriviaSyntax>();
-
-            // Directives contain the EOL, so if the position is within the full span of the
-            // directive, then it is on that line, the only exception is if the directive is on the
-            // last line, the position at the end if technically not contained by the directive but
-            // its also not on a new line, so it should be considered part of the preprocessor
-            // context.
-            if (directive == null)
-            {
-                return false;
-            }
-
-            return
-                directive.FullSpan.Contains(position) ||
-                directive.FullSpan.End == syntaxTree.GetRoot(cancellationToken).FullSpan.End;
-        }
-
         public static bool IsPreProcessorDirectiveContext(this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
         {
             var leftToken = syntaxTree.FindTokenOnLeftOfPosition(position, cancellationToken, includeDirectives: true);
